componentized the demo Button
