+++ client/src/helpers/DemoUserCredential/DemoUserCredential.ts
@@ -1,0 +1,4 @@
+export const demoUserCredential = {
+  email: 'johndoe@gmail.com',
+  password: 'johndoe',
+};
